refactor: move performQualityAnalysis out of the message-listener module
