+++ src/content/content.ts
@@ -1,14 +1,10 @@
 // Content script for WebQualityAnalyzer extension
 import { browser } from '../shared/browser';
-import { analyzeAccessibility } from './analyzers/accessibility';
-import { analyzeSEO } from './analyzers/seo';
-import { analyzePerformance } from './analyzers/performance';
-import type { AnalysisResult, CategoryResult } from './types';
+import { performQualityAnalysis } from './analyze';
 import { AnalyzerSettings, DEFAULT_SETTINGS } from '../shared/settings';
 
 export type { AnalysisResult, CategoryResult, Issue } from './types';
-
-const EMPTY_CATEGORY: CategoryResult = { score: 100, issues: [], suggestions: [] };
+export { performQualityAnalysis } from './analyze';
 
 // Listen for messages from popup
 browser.runtime.onMessage.addListener((request: unknown) => {
@@ -19,29 +15,3 @@ browser.runtime.onMessage.addListener((request: unknown) => {
   }
   return undefined;
 });
-
-export function performQualityAnalysis(
-  settings: AnalyzerSettings = DEFAULT_SETTINGS
-): AnalysisResult {
-  const accessibility = settings.accessibility.enabled
-    ? analyzeAccessibility(settings.accessibility)
-    : EMPTY_CATEGORY;
-  const seo = settings.seo.enabled ? analyzeSEO(settings.seo) : EMPTY_CATEGORY;
-  const performance = settings.performance.enabled
-    ? analyzePerformance(settings.performance)
-    : EMPTY_CATEGORY;
-
-  const overallScore = Math.round(
-    (accessibility.score + seo.score + performance.score) / 3
-  );
-
-  return {
-    score: overallScore,
-    pageInfo: {
-      url: window.location.href,
-      title: document.title,
-      timestamp: new Date().toISOString(),
-    },
-    categories: { accessibility, seo, performance },
-  };
-}
